--- a/app/models/reports/matview.rb
+++ b/app/models/reports/matview.rb
@@ -3,7 +3,7 @@ module Reports
     def self.refresh
       ActiveRecord::Base.transaction do
         ActiveRecord::Base.connection.execute("SET LOCAL TIME ZONE '#{Period::REPORTING_TIME_ZONE}'")
-        Scenic.database.refresh_materialized_view(table_name, concurrently: false, cascade: false)
+        Scenic.database.refresh_materialized_view(table_name, concurrently: true, cascade: false)
       end
     end
   end